break map out into several files
