+++ src/components/Jobs/JobCard.js
@@ -28,7 +28,7 @@ const firstSentence = (text) =>
     : text.substr(0, text.indexOf('.') + 1)
 
 const JobCard = ({ job, children, sourceImagePosition = 'footer' }) =>
-  <Card>
+  console.log(job) || <Card>
     {
       sourceImagePosition === 'header' &&
       <Header>
